Version Two: exported tweet code to its own js file. Initially tried to include it and export it from app.js, but this would not work. Got the current tweets to display on pug template. However, there is an issue with refreshing the page to get the current tweets. Successfully updating the page with new tweets currently requires two refreshes: one to repopulate the underlying array, and another to feed the repopulated array to the pug template. This is because the page is rendering before the array can be repopulated through communication with the API. The only solution I can come up with is to refactor the code such that the asynchronous call to the API occurs in the route file and uses a promise to hold off on rendering until the API call is completed. This will be done in version 3 below.
Version Three: read the explanation above re: promise integration and related refactoring. This worked! Basically changed tweets file to functions.js and brought the API call directly into the index.js file in the routes folder. Then, through use of a simple promise, the Tweets are repopulated and displayed every time the page is displayed (reloads included). Added functionality such that clicking like button will result in like on Twitter page, clicking retweet will result in a retweet and clicking reply will cause a modal window to pop up that allows the user to post a reply to a tweet. Not part of the project requirements, but I thought I’d do it anyway. Also ensured that links are availble to the user in both tweets and replies and that the correct time since tweets were posted is displayed.